feat: add a depends only devshell
Add a depends devshell. This shell only includes packages needed for
doing a depends build and some basic debugging tools.

This needs to be a separate shell because depends building will fail
if there are system packages installed. Conceptually, its also cleaner
to only include the necessary dependencies for a depends build since
this is precisely the scenario in which depends is useful.